Chess Movement Assignment - Giorgio Vallesi
This code is not working. It lacks logic for every piece, it is only able to correctly switch between players and move pieces in empty or opponent cases.
I apologise I am not presenting a working code on time, I have tried my best. I am submitting what I have for now and will try to get a working code in the next days.
Thanks for your attention.