Chess Movement Assignment
Giorgio Vallesi - Chess Movement Assignnment

This code is not working. It lacks logic for every piece, it is only able to correctly switch between players and move pieces in empty or opponent cases.
I apologise I am not presenting a working code on time, I have tried my best. I am submitting what I have for now and will try to get a working code in the next days.
I was working on the queen and pawns movement, I have created a new class MoveGenerator that would have been able to output the possible moves givenn a piece and its location.
Unfortunately, I could not make it on time.
Thanks for your attention.